requirement.txt
astroid==1.6.2
backports.functools-lru-cache==1.5
certifi==2018.1.18
chardet==3.0.4
click==6.7
configparser==3.5.0
enum34==1.1.6
Flask==0.12.2
Flask-JWT==0.3.2
futures==3.2.0
get==0.0.39
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.12
post==0.0.26
public==0.0.65
PyJWT==1.4.2
pylint==1.8.3
PyMySQL==0.8.0
query-string==0.0.28
request==0.0.26
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
